Fix the robot not applying brake on disable
The problem was that VelocityTorqueCurrentFOC requires it to override the neutral mode output to coast instead of brake, since ~0 output triggers neutral mode but VelocityTorqueCurrentFOC pretty easily hits 0 output due to the nature of TorqueCurrentFOC. This neutral mode override carries over into disable.

CTRE is looking into a fix on their end, but for now, explicitly commanding a static brake seems to work well.

- Tada